+++ pkg/runtime/expressions/for.go
@@ -4,21 +4,21 @@ import (
 	"context"
 	"github.com/MontFerret/ferret/pkg/runtime/core"
 	"github.com/MontFerret/ferret/pkg/runtime/expressions/clauses"
-	"github.com/MontFerret/ferret/pkg/runtime/expressions/source"
+	"github.com/MontFerret/ferret/pkg/runtime/expressions/datasource"
 	"github.com/MontFerret/ferret/pkg/runtime/values"
 	"github.com/pkg/errors"
 )
 
 type ForExpression struct {
 	src        core.SourceMap
-	dataSource source.DataSource
+	dataSource datasource.DataSource
 	predicate  core.Expression
 	spread     bool
 }
 
 func NewForExpression(
 	src core.SourceMap,
-	dataSource source.DataSource,
+	dataSource datasource.DataSource,
 	predicate core.Expression,
 	spread bool,
 ) (*ForExpression, error) {
